balance point

+++ 14-binary_tree_balance.c
@@ -8,7 +8,7 @@
 
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	int lh = 0, rh = 0;
+	size_t lh = 0, rh = 0;
 
 	if (!tree)
 		return (0);
